requirements.txt
pandas
numpy
seaborn
matplotlib
scikit-learn
xgboost
joblib
streamlit==1.35.0
plotly
mysql-connector-python
-e .